backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
sqlalchemy
sqlmodel